feat: delete user

+++ apps/api/src/handlers/users.ts
@@ -7,6 +7,7 @@ import {
   getAllUsers,
   getOneUser,
   removeUserRole,
+  softDeleteUser,
   updateUser,
   updateUserRole,
 } from '../routes/user.js';
@@ -24,7 +25,12 @@ users.openapi(getAllUsers, async (c) => {
   await checkRole(roles, false, [Role.ADMIN]);
   const { skip, take } = c.req.valid('query');
   const { from, to } = getPagination(skip, take - 1);
-  const { data, error } = await supabase.from('USERS').select('*, roles:ROLES (id, name)').range(from, to);
+  const { data, error } = await supabase
+    .from('USERS')
+    .select('*, roles:ROLES (id, name)')
+    .range(from, to)
+    .order('created_at', { ascending: true })
+    .filter('deleted_at', 'is', null);
 
   if (error) {
     return c.json({ error: error.message }, 500);
@@ -37,7 +43,12 @@ users.openapi(getOneUser, async (c) => {
   const roles = c.get('user').roles || [];
   await checkRole(roles, true);
   const { id } = c.req.valid('param');
-  const { data, error } = await supabase.from('USERS').select('*, roles:ROLES (id, name)').eq('id', id).single();
+  const { data, error } = await supabase
+    .from('USERS')
+    .select('*, roles:ROLES (id, name)')
+    .eq('id', id)
+    .filter('deleted_at', 'is', null)
+    .single();
 
   if (error || !data) {
     return c.json({ error: 'User not found' }, 404);
@@ -55,6 +66,20 @@ users.openapi(updateUser, async (c) => {
 
   const allowed = [Role.MODERATOR, Role.ADMIN, Role.DIRECTOR];
   if (roles?.some((role) => allowed.includes(role))) {
+    const { data: dataExist, error: errorExist } = await supabase
+      .from('USERS')
+      .select('id, deleted_at')
+      .eq('id', id)
+      .single();
+
+    if (errorExist || !dataExist) {
+      return c.json({ error: 'User not found' }, 404);
+    }
+
+    if (dataExist.deleted_at) {
+      return c.json({ error: 'The user was deleted' }, 400);
+    }
+
     const { data, error } = await supabase
       .from('USERS')
       .update({ first_name, last_name, username })
@@ -90,6 +115,16 @@ users.openapi(addUserRole, async (c) => {
   const roles = c.get('user').roles || [];
   await checkRole(roles, false, [Role.ADMIN]);
 
+  const { data, error: errorID } = await supabase.from('USERS').select('id, deleted_at').eq('id', id).single();
+
+  if (errorID || !data) {
+    return c.json({ error: 'User not found' }, 404);
+  }
+
+  if (data.deleted_at) {
+    return c.json({ error: 'The user was deleted' }, 400);
+  }
+
   const { error } = await supabase.from('USERS_ROLES').insert({ id_user: id, id_role });
 
   if (error) {
@@ -104,6 +139,16 @@ users.openapi(removeUserRole, async (c) => {
   const { id_role } = c.req.valid('json');
   const roles = c.get('user').roles || [];
   await checkRole(roles, false, [Role.ADMIN]);
+
+  const { data, error: errorID } = await supabase.from('USERS').select('id, deleted_at').eq('id', id).single();
+
+  if (errorID || !data) {
+    return c.json({ error: 'User not found' }, 404);
+  }
+
+  if (data.deleted_at) {
+    return c.json({ error: 'The user was deleted' }, 400);
+  }
 
   const { error } = await supabase.from('USERS_ROLES').delete().eq('id_user', id).eq('id_role', id_role);
 
@@ -145,6 +190,16 @@ users.openapi(updateUserRole, async (c) => {
   const userRoles = c.get('user').roles || [];
   await checkRole(userRoles, false, [Role.ADMIN]);
 
+  const { data, error: errorID } = await supabase.from('USERS').select('id, deleted_at').eq('id', id).single();
+
+  if (errorID || !data) {
+    return c.json({ error: 'User not found' }, 404);
+  }
+
+  if (data.deleted_at) {
+    return c.json({ error: 'The user was deleted' }, 400);
+  }
+
   const { error } = await supabase.from('USERS_ROLES').delete().eq('id_user', id);
 
   if (error) {
@@ -163,4 +218,48 @@ users.openapi(updateUserRole, async (c) => {
   }
 
   return c.json({ message: 'Roles updated' }, 200);
+});
+
+users.openapi(softDeleteUser, async (c) => {
+  const { id } = c.req.valid('param');
+  const roles = c.get('user').roles || [];
+  await checkRole(roles, false, [Role.ADMIN]);
+
+  const { data: user_data, error: errorID } = await supabase
+    .from('USERS')
+    .select('id, deleted_at')
+    .eq('id', id)
+    .single();
+
+  if (errorID || !user_data) {
+    return c.json({ error: 'User not found' }, 404);
+  }
+
+  if (user_data.deleted_at) {
+    return c.json({ error: 'User already deleted' }, 400);
+  }
+
+  const { error: errorRole } = await supabase.from('USERS_ROLES').delete().eq('id_user', id);
+
+  if (errorRole) {
+    return c.json({ error: 'Failed to delete user' }, 500);
+  }
+
+  const { error } = await supabase
+    .from('USERS')
+    .update({
+      username: 'Deleted',
+      first_name: 'Deleted',
+      last_name: 'Deleted',
+      email: `Deleted${user_data.id}`,
+      date_validity: null,
+      deleted_at: new Date().toISOString(),
+    })
+    .eq('id', id);
+
+  if (error) {
+    return c.json({ error: 'Failed to delete user' }, 500);
+  }
+
+  return c.json({ message: 'User deleted' }, 200);
 });
